requirements.txt
moviepy 
openai-whisper 
openai